Функционал перенесен в presenter, реализована переключение между попапами с сохранением только одного в открытом состоянии
