+++ src/view/abstract.js
@@ -15,7 +15,9 @@ export default class Abstract {
   }
 
   getElement() {
-    this._element = createElement(this.getTemplate());
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
 
     return this._element;
   }
